+++ components/todo-item.tsx
@@ -1,33 +1,15 @@
 'use client';
 
-import { useDeleteTodo, useToggleTodoStatus } from '@/hooks/todo';
 import { Todo } from '@/types';
-import { useQueryClient } from '@tanstack/react-query';
 import clsx from 'clsx';
+import { TodoStatusCheckbox } from './todo-status-checkbox';
+import { TodoDeleteButton } from './todo-delete-button';
 
 interface TodoProps {
   todo: Todo;
 }
 
 export function TodoItem({ todo }: TodoProps) {
-  const queryClient = useQueryClient();
-
-  const { mutate: toggleTodoStatus, isLoading: isToggleTodoStatusLoading } =
-    useToggleTodoStatus(todo, {
-      onSuccess: () => queryClient.refetchQueries(['todos']),
-    });
-
-  const { mutate: deleteTodo, isLoading: isDeleteTodoLoading } = useDeleteTodo(
-    todo.id,
-    {
-      onSuccess: () => queryClient.refetchQueries(['todos']),
-    }
-  );
-
-  const handleCompleteCheckboxChange = () => toggleTodoStatus();
-
-  const handleDeleteButtonClick = () => deleteTodo();
-
   return (
     <article
       id={`todo-${todo.id}`}
@@ -35,20 +17,8 @@ export function TodoItem({ todo }: TodoProps) {
     >
       <span>{todo.id}</span>
       <span>{todo.task}</span>
-      <input
-        type="checkbox"
-        checked={todo.is_complete || false}
-        onChange={handleCompleteCheckboxChange}
-        disabled={isToggleTodoStatusLoading}
-      />
-      <button
-        type="button"
-        className="text-red-500"
-        onClick={handleDeleteButtonClick}
-        disabled={isDeleteTodoLoading}
-      >
-        (X)
-      </button>
+      <TodoStatusCheckbox todo={todo} />
+      <TodoDeleteButton todoId={todo.id} />
     </article>
   );
 }
